refactor: improve project structure following fastapi best practices
