+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.12.0
 WORKDIR /jvtips-backend
-COPY ./app /jvtips-backend/app
+COPY ./src /jvtips-backend/src
 COPY requirements.txt /jvtips-backend/requirements.txt
 COPY .env /jvtips-backend/.env
 RUN pip install -r requirements.txt
