+++ docker-compose.yml
@@ -1,9 +1,9 @@
 services:
   api:
     build: .
-    command: uvicorn app.main:app --host 0.0.0.0 --port 3000
+    command: uvicorn src.main:app --host 0.0.0.0 --port 3000
     volumes:
-      - .:/app
+      - .:/src
     ports:
       - 3000:3000
     depends_on:
